Connect ToolComponent to editor state
