+++ ui/src/components/editor/PlateSidebar/Tools/ApplyTool/ToolComponent/connect.js
@@ -1,0 +1,23 @@
+import { connect } from 'react-redux';
+
+import { ToolComponent } from './ToolComponent';
+import {
+  selectToolComponents,
+  deselectToolComponents,
+  removeToolComponents,
+  addTimepointToComponent,
+  updateTimepoint,
+  deleteTimepoint,
+} from '../../../../../../store/editorActions';
+
+const mapDispatch = {
+  onSelect: selectToolComponents,
+  onDeselect: deselectToolComponents,
+  onRemoveClick: removeToolComponents,
+  onAddTimepointClick: addTimepointToComponent,
+  onTimepointChange: updateTimepoint,
+  onTimepointDeleteClick: deleteTimepoint,
+};
+
+const connected = connect(null, mapDispatch)(ToolComponent);
+export { connected as ToolComponent };
