+++ ui/src/components/editor/PlateSidebar/Tools/ApplyTool/ToolComponent/index.js
@@ -1,0 +1,2 @@
+import { ToolComponent } from './connect';
+export { ToolComponent };
