+++ ui/src/store/designExperiment.js
@@ -1,12 +1,10 @@
 import { createSlice } from 'redux-starter-kit';
-import validate from 'validate.js';
 
 import {
   STATUS_DRAFT,
   STATUS_COMPLETED,
   DEFAULT_COMPONENT_COLOR_CODES,
 } from '../constants';
-import { createTimepoint } from './plateFunctions';
 
 const designExperiment = createSlice({
   slice: 'designExperiment',
@@ -54,74 +52,6 @@ const designExperiment = createSlice({
       state.experiment = experiment;
       state.plateSize = plateSize;
       state.steps.stepOneCompleted = true;
-    },
-    removeToolComponents(state, action) {
-      const componentsToRemove = action.payload.components;
-      const { toolComponents } = state;
-      const idsToRemove = componentsToRemove.map(component => component.id);
-      idsToRemove.forEach(id => {
-        const index = toolComponents.findIndex(
-          component => component.id === id
-        );
-        if (index > -1) {
-          toolComponents.splice(index, 1);
-        }
-      });
-    },
-    selectToolComponents(state, action) {
-      const { components } = action.payload;
-      components.forEach(component => {
-        const stateComponent = getToolComponentFromState(component.id, state);
-        stateComponent.selected = true;
-      });
-    },
-    deselectToolComponents(state, action) {
-      const { components } = action.payload;
-      components.forEach(component => {
-        const stateComponent = getToolComponentFromState(component.id, state);
-        stateComponent.selected = false;
-      });
-    },
-    addTimepointToComponent(state, action) {
-      const { component } = action.payload;
-      const stateComponent = getToolComponentFromState(component.id, state);
-      const { timepoints } = stateComponent;
-      let time;
-      if (timepoints.length > 0) {
-        const max = timepoints.reduce((highest, current) => {
-          return current.time > highest ? current.time : highest;
-        }, 0);
-        time = max + 24;
-      }
-      timepoints.push(createTimepoint(component.type, time));
-    },
-    updateTimepoint(state, action) {
-      const { component, name, value, index } = action.payload;
-      const stateComponent = getToolComponentFromState(component.id, state);
-      const { timepoints } = stateComponent;
-      const timepoint = timepoints[index];
-      timepoint[name] = value;
-      const errors = validate.single(
-        timepoints,
-        { timepoints: true },
-        { fullMessages: false }
-      );
-      if (!errors) {
-        stateComponent.isValid = true;
-        stateComponent.errors = [];
-        state.toolComponentsValid = true;
-      } else {
-        stateComponent.isValid = false;
-        stateComponent.errors = errors;
-        state.toolComponentsValid = false;
-      }
-    },
-    deleteTimepoint(state, action) {
-      const { component, index } = action.payload;
-      if (index > 0) {
-        const stateComponent = getToolComponentFromState(component.id, state);
-        stateComponent.timepoints.splice(index, 1);
-      }
     },
     toggleComponentEditing(state, action) {
       const { component } = action.payload;
